@@ -17,7 +17,7 @@
     "gulp": "^3.9.1",
     "gulp-typescript": "^3.0.2",
     "prettydiff": "^99.0.1",
-    "typescript": "^2.0.3",
+    "typescript": "^2.1.0-dev.20161008",
     "typings": "^1.4.0"
   }
 }
