@@ -1,23 +1,16 @@
-// InputButton.js
 /// <reference path="../.vscode/typings/react-native/react-native.d.ts"/>
 
-
 import React, { Component } from "react";
-import {
-    Text,
-    TouchableHighlight,
-} from "react-native";
+import {Text, TouchableHighlight} from "react-native";
 import Style from "./Style";
 
-// Typescript Interfaces
 interface InpProps {
-    highlight: any;
-    onPress: any;
-    value: any;
+    highlight: boolean;
+    onPress: () => void;
+    value: string | number;
 }
 
 export default class InputButton extends Component<InpProps, {}> {
-
     render() {
         return (
             <TouchableHighlight style={[Style.inputButton, this.props.highlight ? Style.inputButtonHighlighted : null]}
